pyproject.toml
[project]
name = "sentrysearch"
version = "0.2.0"
description = "Search dashcam footage using natural language queries"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "python-dotenv>=1.0",
    "google-genai>=1.0",
    "imageio-ffmpeg>=0.5",
    "protobuf>=4.0,<7",
    "psycopg[binary]>=3.1",
    "pgvector>=0.3",
    "boto3>=1.34",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]

[project.optional-dependencies]
tesla = ["geopy"]

[project.scripts]
sentrysearch = "sentrysearch.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentrysearch"]

[dependency-groups]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.uv]
constraint-dependencies = [
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
    "opentelemetry-proto>=1.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["sentrysearch"]
omit = ["sentrysearch/dashcam_pb2.py", "sentrysearch/local_embedder.py"]

[tool.coverage.report]
show_missing = true
